# paddock/data/apps/claims-api/app/api/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for DecisionDecision.
const (
	Approve DecisionDecision = "approve"
	Deny    DecisionDecision = "deny"
)

// Claim defines model for Claim.
type Claim struct {
	AmountCents  int     `json:"amount_cents"`
	DecisionNote *string `json:"decision_note,omitempty"`
	Description  string  `json:"description"`
	HolderUid    string  `json:"holder_uid"`
	Id           string  `json:"id"`
	IncidentDate string  `json:"incident_date"`
	PolicyNumber string  `json:"policy_number"`
	Status       string  `json:"status"`
	Version      int     `json:"version"`
}

// ClaimEnvelope defines model for ClaimEnvelope.
type ClaimEnvelope struct {
	Claim Claim `json:"claim"`
}

// ClaimList defines model for ClaimList.
type ClaimList struct {
	Claims []Claim `json:"claims"`
}

// ClaimSubmission defines model for ClaimSubmission.
type ClaimSubmission struct {
	AmountCents  int    `json:"amount_cents"`
	Description  string `json:"description"`
	IncidentDate string `json:"incident_date"`
	PolicyNumber string `json:"policy_number"`
}

// Decision defines model for Decision.
type Decision struct {
	Decision DecisionDecision `json:"decision"`
	Note     *string          `json:"note,omitempty"`
	Version  int              `json:"version"`
}

// DecisionDecision defines model for Decision.Decision.
type DecisionDecision string

// FieldErrors defines model for FieldErrors.
type FieldErrors struct {
	Errors map[string]string `json:"errors"`
}

// Health defines model for Health.
type Health struct {
	Status string `json:"status"`
}

// Problem defines model for Problem.
type Problem struct {
	Detail *string `json:"detail,omitempty"`
	Title  string  `json:"title"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Problem

// Conflict defines model for Conflict.
type Conflict = Problem

// Forbidden defines model for Forbidden.
type Forbidden = Problem

// NotFound defines model for NotFound.
type NotFound = Problem

// Unauthorized defines model for Unauthorized.
type Unauthorized = Problem

// Unprocessable defines model for Unprocessable.
type Unprocessable = FieldErrors

// SubmitClaimJSONRequestBody defines body for SubmitClaim for application/json ContentType.
type SubmitClaimJSONRequestBody = ClaimSubmission

// DecideClaimJSONRequestBody defines body for DecideClaim for application/json ContentType.
type DecideClaimJSONRequestBody = Decision